NAV-22905: Flytter standard forskyvningslogikk for vilkår ut i egen metode og fil
